move groupedMetric to legacy now that we have pivot!
