+++ src/example-types/legacy/groupedMetric.js
